{{cookiecutter.project_slug}}/requirements/local.txt
-r ./base.txt

ipdb==0.13.4
Sphinx==3.4.3
Werkzeug==1.0.1

# Testing
# ------------------------------------------------------------------------------
factory-boy==3.2.0
mock==4.0.3
mypy==0.800
pytest<6.0.0,>=3.3.0
pytest-sugar==0.9.4

# Code quality
# ------------------------------------------------------------------------------
black==20.8b1
coverage==5.3.1
flake8==3.8.4
{%- if cookiecutter.use_celery == 'y' %}
pylint-celery==0.3
{%- endif %}
pylint-django==2.4.2

# Django
# ------------------------------------------------------------------------------
django-coverage-plugin==1.8.0
django-debug-toolbar==3.2
pytest-django==3.5.1
